Rename useURLSyncedString to useUrlSyncedString due to eslint complains
